added multi-server connections per client
